Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "bcrypt>=4.3.0",
    "chromadb>=1.0.17",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.6",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.1",
    "lark>=1.2.2",
    "motor>=3.7.1",
    "pyjwt>=2.10.1",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]
